[#12235] Improve windowRangeCount for TimeWindows to be precomputed

+++ metric-module/metric-commons/src/main/java/com/navercorp/pinpoint/metric/common/util/TimeUtils.java
@@ -27,7 +27,7 @@ import java.util.List;
 public class TimeUtils {
 
     public static List<Long> createTimeStampList(TimeWindow timeWindow) {
-        List<Long> timestampList = new ArrayList<>((int) timeWindow.getWindowRangeCount());
+        List<Long> timestampList = new ArrayList<>(timeWindow.getWindowRangeCount());
 
         for (Long timestamp : timeWindow) {
             timestampList.add(timestamp);
